public: update filesystem to resolve crashes
found it while trying to get the shaderapidx9.dll compatible with gmod

+++ public/filesystem.h
@@ -517,10 +517,10 @@ public:
 	virtual void			Close( FileHandle_t file ) = 0;
 
 
-	virtual void			Seek( FileHandle_t file, int pos, FileSystemSeek_t seekType ) = 0;
-	virtual unsigned int	Tell( FileHandle_t file ) = 0;
-	virtual unsigned int	Size( FileHandle_t file ) = 0;
-	virtual unsigned int	Size( const char *pFileName, const char *pPathID = 0 ) = 0;
+	virtual void				Seek( FileHandle_t file, long long pos, FileSystemSeek_t seekType ) = 0;
+	virtual unsigned long long	Tell( FileHandle_t file ) = 0;
+	virtual unsigned long long	Size( FileHandle_t file ) = 0;
+	virtual unsigned long long	Size( const char *pFileName, const char *pPathID = 0 ) = 0;
 
 	virtual void			Flush( FileHandle_t file ) = 0;
 	virtual bool			Precache( const char *pFileName, const char *pPathID = 0 ) = 0;
